+++ packages/core/src/languages/html/index.ts
@@ -53,7 +53,6 @@ export type HtmlLanguagePluginOptions = {
  *     new HtmlLanguagePlugin(),
  *   ],
  * });
- *
  * @example
  * webmangler({
  *   plugins: [
@@ -65,7 +64,6 @@ export type HtmlLanguagePluginOptions = {
  *     }),
  *   ],
  * });
- *
  * @since v0.1.0
  * @version v0.1.21
  */